# Optimize NSString.lineAndCharacterForByteOffset()