Add `NSString.CacheContainer.lineAndCharacterForByteOffset(_:)`

This commit reduces the duration of linting KeychainAccess by SwiftLint (8d529f1) from:
```
swiftlint lint  27.27s user 0.25s system 96% cpu 28.424 total
```
to:
```
swiftlint lint  3.76s user 0.10s system 89% cpu 4.291 total
```

This is that I mentioned in #127