@jpsim 2015-12-25
Ouch, this was a heavy one. I'd like to integrate some really rough estimate performance tests, but with tests running on a wide variety of systems, I'm not sure it's possible to do reliably. Do you have any ideas?