@norio-nomura 2015-12-25
Testing on CI? I don't have enough knowledge about that.
I know that Xcode has performance test feature triggering by performance regressions. But that needs setting base performance thresholds on real machine and may not fit CI services.